#38 Utils refactor
