+++ src/main/java/ua/com/juja/microservices/teams/slackbot/util/Utils.java
@@ -2,9 +2,6 @@ package ua.com.juja.microservices.teams.slackbot.util;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.web.client.HttpClientErrorException;
 import ua.com.juja.microservices.teams.slackbot.exceptions.ApiError;
 
 import java.io.IOException;
@@ -25,25 +22,28 @@ public class Utils {
         }
     }
 
-    public static HttpHeaders setupJsonHttpHeaders() {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON_UTF8);
-        return headers;
-    }
-
-    public static ApiError convertToApiError(HttpClientErrorException httpClientErrorException) {
-        ObjectMapper mapper = new ObjectMapper();
-        try {
-            return mapper.readValue(httpClientErrorException.getResponseBodyAsString(), ApiError.class);
-        } catch (IOException e) {
-            return new ApiError(
-                    500, "BotInternalError",
-                    "I'm, sorry. I cannot parse api error message from remote service :(",
-                    "Cannot parse api error message from remote service",
-                    e.getMessage(),
-                    Collections.singletonList(httpClientErrorException.getMessage())
-            );
+    public static ApiError convertToApiError(String message) {
+        int contentExists = message.indexOf("content:");
+        ApiError apiError = new ApiError(
+                500, "BotInternalError",
+                "I'm, sorry. I cannot parse api error message from remote service :(",
+                "Cannot parse api error message from remote service",
+                message, Collections.singletonList(message));
+        if (contentExists != -1) {
+            String apiMessage = message.substring(contentExists + 8);
+            ObjectMapper mapper = new ObjectMapper();
+            try {
+                apiError = mapper.readValue(apiMessage, ApiError.class);
+            } catch (IOException e) {
+                apiError = new ApiError(
+                        500, "BotInternalError",
+                        "I'm, sorry. I cannot parse api error message from remote service :(",
+                        "Cannot parse api error message from remote service",
+                        e.getMessage(), Collections.singletonList(message)
+                );
+            }
         }
+        return apiError;
     }
 
     private static String listToStringWithDelimeter(Set<String> list, String delimeter) {
